Issue: User can only edit their own posts (#12)
Added simple navbar styling
Navbar uses routing instead of href

+++ client/src/app/posts/post-details/post-details.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Post } from 'src/app/models/post.model';
 import { PostService } from '../post.service';
 import { ActivatedRoute, Router } from '@angular/router'
+import { AuthService } from 'src/app/authentication/auth.service';
 
 @Component({
   selector: 'app-post-details',
@@ -11,12 +12,16 @@ import { ActivatedRoute, Router } from '@angular/router'
 export class PostDetailsComponent implements OnInit {
 
   post: Post = new Post();
+  userId: number = 0;
 
-  constructor(private postService: PostService, private activatedRoute: ActivatedRoute, private router: Router) { }
+  constructor(private postService: PostService, private activatedRoute: ActivatedRoute, private router: Router, private authService: AuthService) { }
 
   ngOnInit(): void {
     // Gets id parameter from URL
     const id = this.activatedRoute.snapshot.params["id"];
+
+    // Gets ID from current user
+    this.userId = this.authService.getCurrentUser().getUserId;
 
     // Gets Post with id from database
     this.postService.getPost(id).then(post => {
@@ -25,6 +30,7 @@ export class PostDetailsComponent implements OnInit {
       console.log(error);
     });
   }
+  
   /**
    * Moves to edit page
    */
@@ -36,11 +42,13 @@ export class PostDetailsComponent implements OnInit {
    * Deletes post in database and navigates to post list
    */
   deletePost() {
-    this.postService.deletePost(this.post.getId).then(data => {
-      console.log("Successfully deleted post: " + this.post.getId);
-      this.router.navigateByUrl("/annonse");
-    }).catch(error => {
-      console.log(error);
-    });
+    if (this.userId == this.post.getOwner) {
+      this.postService.deletePost(this.post.getId).then(data => {
+        console.log("Successfully deleted post: " + this.post.getId);
+        this.router.navigateByUrl("/annonse");
+      }).catch(error => {
+        console.log(error);
+      });
+    }
   }
 }
